Solved some problems with tasks names not being compatible with old generated sources

+++ androidGenerator.rb
@@ -66,7 +66,7 @@ class AndroidGenerator
       res=Mustache.render(File.open("templates/android/"+aVersion+"/android_base_service.mustache").read,parameters)
       File.open(logicBaseFile, 'w') { |file| file.write(res) }
       res=Mustache.render(File.open("templates/android/"+aVersion+"/android_service.mustache").read,parameters)
-      File.open(logicFile, 'w') { |file| file.write(res) }
+      File.open(logicFile, 'w') { |file| file.write(res) }      unless File.exists?(logicFile)
     end
     ########## Tasks
     puts 'Tasks'
@@ -95,6 +95,6 @@ class AndroidGenerator
     ############ Copy common folder
     puts 'Copying common folder'
     FileUtils.cp_r Dir.glob("templates/android/"+aVersion+"/static/*"),aOutput
-    
+
   end
 end
